Refactor mobs to be in combat
